create and display process buffer like other buffers
Previously such buffers were created using `magit-process-buffer' which
does not use `magit-mode-setup' (nor `magit-mode-display-buffer' which
is used by the latter).  That meant that some information which is
usually associated with Magit buffers, was not for such buffers.

Now `magit-process-buffer' uses `magit-mode-setup', and it replaces both
the old `magit-process-buffer' as well as the command `magit-process'.

This also fixes an issue.  Previously an error during the initial
creation of the process caused `magit-process-truncate-log' to always
fail from then on.