Use the "noglob" option on Cygwin and MSYS
By default, Cygwin and MSYS executables attempt to expand globs when
called by a program other than a Unix shell.  The intent is to provide
globbing when, for instance, using Cygwin's grep from the Windows
command prompt.

However, this causes problems when calling the Cygwin or MSYS git from
Magit in a native Windows Emacs.  The solution is to add "noglob" to the
"CYGWIN" and "MSYS" environment variables to disable this behavior.

See https://cygwin.com/cygwin-ug-net/using-cygwinenv.html

Fixes #2246